LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/2-iframes/parent-no-child1-no-subdomain-child2-yes-subdomain.sub.https-expected.txt


PASS Parent to child1: setting document.domain must give sync access
PASS Parent to child2: setting document.domain must give sync access
PASS child1 to child2: setting document.domain must give sync access
PASS child2 to child1: setting document.domain must give sync access
FAIL parent: originAgentCluster must equal false assert_equals: expected (boolean) false but got (undefined) undefined
FAIL child1: originAgentCluster must equal false assert_equals: expected (boolean) false but got (undefined) undefined
FAIL child2: originAgentCluster must equal false assert_equals: expected (boolean) false but got (undefined) undefined
